+++ composeApp/src/jvmMain/kotlin/io/github/kdroidfilter/seforimapp/features/screens/bookcontent/BookContentScreen.kt
@@ -12,7 +12,6 @@ fun BookContentScreen() {
     MainBookContentLayout(
         uiState = uiState,
         linesPagingData = viewModel.linesPagingData, // Pass the lines paging data flow
-        commentsPagingData = viewModel.commentsPagingData, // Pass the comments paging data flow
         buildCommentariesPagerFor = viewModel::buildCommentariesPagerFor,
         getAvailableCommentatorsForLine = viewModel::getAvailableCommentatorsForLine,
         onEvent = viewModel::onEvent
